sinkhorn_clip_demo/requirements.txt
torch>=2.0
pytest>=8.0
transformers>=4.40
pillow>=10.0
matplotlib>=3.8
numpy>=1.26
